moved personalaccesskey prompts into cms-cli and removing open and inquirer dep from cms-lib
