+++ packages/cms-cli/commands/init.js
@@ -16,11 +16,8 @@ const { logger } = require('@hubspot/cms-lib/logger');
 const {
   updateConfigWithPersonalAccessKey,
 } = require('@hubspot/cms-lib/personalAccessKey');
-const {
-  promptUser,
-  personalAccessKeyPrompt,
-  PORTAL_NAME,
-} = require('@hubspot/cms-lib/lib/prompts');
+const { PORTAL_NAME } = require('@hubspot/cms-lib/lib/prompts');
+const { promptUser, personalAccessKeyPrompt } = require('../lib/prompts');
 const {
   trackCommandUsage,
   addHelpUsageTracking,
