+++ packages/cms-lib/package.json
@@ -19,7 +19,6 @@
     "ignore": "^5.1.4",
     "js-yaml": "^3.12.2",
     "moment": "^2.24.0",
-    "open": "^7.0.3",
     "p-queue": "^6.0.2",
     "prettier": "^1.19.1",
     "request": "^2.87.0",
